+++ .codeclimate.yml
@@ -11,7 +11,7 @@ checks:
   file-lines:
     enabled: true
     config:
-      threshold: 400
+      threshold: 500
   method-complexity:
     enabled: true
     config:
@@ -29,9 +29,7 @@ checks:
     config:
       threshold: 4
   return-statements:
-    enabled: true
-    config:
-      threshold: 4
+    enabled: false
   similar-code:
     enabled: true
     config:
